environment/software: Save kernel logs
The kernel logs could be useful to understand what was the system
context before a run.

This patch is saving the kernel logs from journalctl into the
'kernel-logs' file.

The json output format is used to ease a future parsing at the cost of a
bigger file.

This operation could take time regarding the size of the kernel journal
but as hwbench will run in the netbooted image, it must be not too big.

Signed-off-by: Erwan Velu <e.velu@criteo.com>