environment: Adding NUMA parsing
This commit is about adding the parsing of the NUMA topology.

numactl -H provides the list of numa domains and the associated cores
into it.

This commit is adding the parsing and a couple of helpers.

Sample output from single and multi numa domains configs are added in
tests. They were used to develop code in TDD.

Numa distances are not yet reported in helpers.

IRA: HAL-1801
Signed-off-by: Erwan Velu <e.velu@criteo.com>