benchmark: Add common parameter hardware
In the future, we might want to give the benchmarks even more
parameters, like access to the environment to evaluate if the benchmark
is runnable or not (example: is this CPU instruction supported ?). For
now, we only pass hardware, but we might pass the software environment
too later.

But we also still want to be able to test the benchmarks parsing, and
evaluation functions. To do that, we factor common parameters in a class
that will be used as argument. We only use interfaces if we can, meaning
that all the interface of the parameters can be reimplemented in mock
objects.

We provide such an example in the test_parse test. In addition, we pass
both software and hardware environments to benchmarks to prepare for
future changes.