engines/stressng: add back qsort module test
We already have the test vectors, they just need a bit of updating. We
only test the output parsing, not the version, which is done in a
separate test for the engine.

We add the new pytest-subtests dependency to be able to differentiate
various loop iterations when we have an error.