add pyright type checker to CI
This adds a checking step that all type annotations are correct. These
are the same checks that should be integarted by default in vscode, or
to anyone using the pyright LSP.

All errors should have been fixed prior to this commit.

Note that pyright is written in typescript, so it relies on having
node.js.
The wrapper used in pypi ("pyright" package) uses the nodeenv python
library to download a node pre-built binary; we have fixed its version
to ensure stability, and do not use system node because it is currently
too old on CI (10.x).