benchmark: Implementing default hosting_cpu_cores=none
The current code was considering hosting_cpu_cores=1 as a default value.

This is an issue since it means using "Core 1" if no directive is given
in the configuration file.

This patch is about defaulting hosting_cpu_cores to "none" which implies
no pinning at all during the bench. It means its up to the OS scheduler
to distribute the workload and processes can migrate during the run
(impacting performance).

To ensure a proper stdout reporting, the stressor message print for each
run is moved inside ExternalBench().

Signed-off-by: Erwan Velu <e.velu@criteo.com>