benchmarks: List jobs, benchmarks and runtime at start time
When starting hwbench with a configuration file, it could be difficult
to estimate the number of benchmarks to be run and the amount of time it
will take.

This patch is about adding two Benchmarks methods to count the overall
runtime and the number of jobs detected in the configuration file.

These methods are also verified in the tests to ensure the parsing did
the right job. To ease the computation and human verification, runtime
of the sample.ini is reduced to 1 second.

A typical output looks like:
	[root@criteo-conformity hwbench]#  python3 -m hwbench.hwbench -c qsort_all.conf
	hwbench: 1 jobs, 9 benchmarks, ETA 225 seconds
	[qsort] stressng/qsort/qsort:   1 stressor for 25s
	[qsort] stressng/qsort/qsort:   2 stressor for 25s
	[qsort] stressng/qsort/qsort:   4 stressor for 25s
	[qsort] stressng/qsort/qsort:   8 stressor for 25s
	[qsort] stressng/qsort/qsort:  16 stressor for 25s
	[qsort] stressng/qsort/qsort:  32 stressor for 25s
	[qsort] stressng/qsort/qsort:  64 stressor for 25s
	[qsort] stressng/qsort/qsort:  96 stressor for 25s
	[qsort] stressng/qsort/qsort: 128 stressor for 25s

Signed-off-by: Erwan Velu <e.velu@criteo.com>