config: Implement groups in hosting_cpu_cores
This commit add the group syntax for hosting_cpu_cores.
Groups are split by a space like "0,64 1,65"

Each group has the same syntax as previous :
- simple integer: <x>
- list of integers: x,y,z
- range of integers: x-z

Each group will be considered as a consistent set of cores during a run.

To ease the implementation, a get_taskset() method is added in
ExternalBench() to return the proper taskset arguments. This method
takes in charge the various notations of get_pinned_cpu().
It also simplify engine's code.

Associated unit tests are added to ensure a proper parsing.

Signed-off-by: Erwan Velu <e.velu@criteo.com>