config: Implement stressor_range=auto
If hosting_cpu_cores is used, setting stressor_range to auto means the
stressor_range will be live adjusted to match the number of cores
selected by hosting_cpu_cores.

A typical usage is like :
	stressor_range=auto
	hosting_cpu_cores=0,64 1,65 0,1,64,65

In this example, three runs will be done with 2 benchmarks instances for
the two first ones and 4 benchmarks for the last one.

This commit adds the associates tests too.


Signed-off-by: Erwan Velu <e.velu@criteo.com>