csv: add tool to convert results.json to csv
For now it prints the CSV on the stdout.

It does not yet convert report-power output. It does not support stream
benchmark output and other cpu benchmarks.

It removes cpu_pin that is a list and contains commas — sidestepping the
csv quoting issue.

Benchmark results are sorted according to job name.