benchmarks: Report ETA in h:m:s
The runtime was reported in seconds but long jobs are generating large
numbers which are a bit difficult to get quickly.

This patch is about reporting the ETA in hours/minutes/seconds.

Before:
	hwbench: 1 jobs, 36 benchmarks, ETA 1260 seconds

After:
	hwbench: 1 jobs, 36 benchmarks, ETA 0h 21m 00s

This commit also ensure the default time is set as a string as all other
config file parameters.

Signed-off-by: Erwan Velu <e.velu@criteo.com>