config: Implement NUMA directive
This commit is about adding the NUMA keyword of hosting cpu cores.

The numa directive uses the same int,list,range syntax as other
configuration directives.

Some sample syntax can be used :
 integer : NUMA0
 range   : NUMA0-7
 list    : NUMA0,1,5

If the request NUMA domain does not exist, a specific error message is
reported.

The config/numa.conf sample file is used to test some of these combo
during the tests.

As per the test_numa() test, NUMA, CPU_INFO and
CPU_CORES instances are created some the sample output and passed to a
MockedCPU object and added to a TestHardware object.
Adding new tests would as easy as changing the sample files.

This commit also pass the hardware object to Config(). This is mandatory
as the NUMA keyword expanding requires the list of cpu cores and numa
topology to be available.

Signed-off-by: Erwan Velu <e.velu@criteo.com>