config: Adding 'simple' helper
This hosting cpu cores helper is implemented to perform the following
scaling.

The hosting_cpu_cores=simple is replaced by a list of physical cores
(and their sibblings) wit the following logic : 1, 2, 3, 4, 8, 16 then +16 up to the core count

For a 64 cores system, this 'simple' helper is replaced by a serie of logical cores count : [2, 4, 6, 8, 16, 32, 64, 96, 128].
It's similar to "0,64 0-1,64-65 0-2,64-66 0-3,64-67 0-7,64-71 0-15,64-79 0-31,64-95 0-47,64-111 0-127"

The 32 and 64 cores config files are replaced by a single and generic one.

Signed-off-by: Erwan Velu <e.velu@criteo.com>