tests: Use 'should_be_fatal()' instead of 'with self.assertRaises(SystemExit):'
The current testing code was doing calls like :

	with self.assertRaises(SystemExit):
		x()
                y()
                z()

The issue with this syntax is that if one of the three functions trigger
the SystemExit, the test passes.

This commit is about implementing a should_be_fatal() function to be
called for every function we want to validate. With this change, the
previous example will be changed by:

	for func in [x, y, z]:
  	    should_be_fatal(func)

Please note we only pass the function name, so this is 'func' and not 'func()'.

This change has been populated to all tests including config/test_parse.

Enabling this change reported a couple of mis-detected issues that got
fixed too in this commit.

Signed-off-by: Erwan Velu <e.velu@criteo.com>