benchmark: Do not validate parameters in constructor
The current code is validating the parameters in the constructor.
This makes the testing code is bit complicated to load benches and
detect a specific failure for a job parameter validation.

This code is moving the validation in validate_jobs_parameters() and
call it from parse_config() which keeps the same default behavior.

parse_config() as a specific argument to ask not calling the validation
to perform it later (the test case).

Signed-off-by: Erwan Velu <e.velu@criteo.com>