bench: Add spike engine
This commit is about adding a new spike engine.

This engine is using engine_module_parameter_base to defines the spiking
timings as per :

	engine_module_parameter_base=high:10 low:5

In this example, each cycle is made of 10 seconds with a high load and 5 seconds idle.
This will be repeated within the runtime period.

When loading the configuration, a check is done to ensure cycles
(low+high) is a modulo of job's runtime. This helps having balanced
cycles during a single job.

The engine is calling stress-ng matrixprod test which is known to generate a
great power consumption. The engine makes a direct call to the stress-ng
during the high time to avoid generating specific jobs for each part of
the cycle. The spike engine will generate a single job doing high/low
cycles during the runtime period.

This benchmark is not generating performance results as we are
considering here only the thermal effects of the spiking.

This is a first implementation of the engine but further modifications
in hwbench will be needed to get a efficient reporting:
- merge report_power inside hwbench to avoid instanciate this tool at
  each run
- collecting the raw values of thermal/power events to study what was
  the cooling response of a particular spike.

Signed-off-by: Erwan Velu <e.velu@criteo.com>