benchmark: Add pre() and post_run()
If some inherited objects needs to override the run() method, we'd love
to keep some part of it as the pre and post parts.

This commit is moving everything before the run into pre_run() and
everything after the run in post_run().

This way, inherited objects can super seed the run but still call pre
and post run helpers.

Signed-off-by: Erwan Velu <e.velu@criteo.com>