hwbench: Improve spike test
The spike test had two main issues:
- it was testing a single core at each iteration
- the output didn't contained the job context

This commit is doing the following:
- adding stressor_range=auto to ensure each iteration is using the right
  number of cores

- Use self.parameters.get_result_format() when calling the post_run :
The spike engine is a bit special since it makes loops of benchmark runs
+ pauses in a row during the job runtime. This behavior prevents this
engine from using the same helpers.
One of big difference is that spike calls directly the post_run() to add
the monitoring metrics in its results.
The issue here is that post_run() was called with an empty argument
meaning only the monitoring metrics are present in the final results.
This commit makes an implicit call to
self.parameters.get_result_format() while its implicit in other engines.
This ensure the output metrics contains both job context _and_
monitoring.
That's a very specific implementation of this engine, this should not
occur often ...

Signed-off-by: Erwan Velu <e.velu@criteo.com>