compgraph: Update get_metric_mean() logic
get_metric_mean() was meant to be using only the specified metric set in
the constructor.

It will be needed to extract other metrics, so let's put the metric_name
a possible option with the same default value as today.

This commit also add get_metric_mean_by_name() to allow providing a
benchmark name instead of the bench object.

Signed-off-by: Erwan Velu <e.velu@criteo.com>