compgraph: Update show_source_file logic in Graph constructor
Graph constructor :
- required trace as a mandatory argument
- show_source_file was True by default

But show_source_file was the only part of the code using trace argument
meaning that if show_source_file is False, trace is provided for
nothing.

This commit is about:
- disabling show_source_file by default
- show_source_file=trace to enable the feature

As a single graph type is enabling show_source_file, this commit makes
the code easier to read and constructor is shorter.

Signed-off-by: Erwan Velu <e.velu@criteo.com>