hwgraph: Add servers chassis power-consumption on enclosure graph
When --same-enclosure feature is used, a graph is generated to compare
the enclosre power consumption versus the sum of servers.

This patch is adding the power consumption of each server on the same
graph. This gives a better understanding of the contribution of each
server to the sum line over time.

Signed-off-by: Erwan Velu <e.velu@criteo.com>