hwgraph: Using serial number if no logical name given
When calling hwgraph with --traces argument, each trace must provide :
- a filename
- a logical name
- a reference power metric

This commit is about considering that if no logical name is given, we
default this field to the serial number included in the trace.

This feature is really useful when plotting chassis at once.

This commit rename "name" to "logical_name" as per the definition above
to get more clarity in the code.

Signed-off-by: Erwan Velu <e.velu@criteo.com>