hwbench: Adding chassis information in dmi dump
When graphing results of multi-node systems, it would be useful to get
the chassis information. This information will be used to generate
specific graphs if all nodes are from the same chassis.

The chassis name is located in chassis_version field which isn't very
obvious. This commit name this field chassis/product which is closer to
what we have for the system.

Signed-off-by: Erwan Velu <e.velu@criteo.com>