hwgraph: Improving memory management
The current code was barely not able to render multiple large traces
because of the excessive memory consumption.

Part of the code was using a workaround to trigger, or not, plot.close("all"), while not really being able to explain why.
It appear that the rendering is done in background making plt.close("all") closing unrendered jobs generating empty images.
A side effect of this, not calling the close() directive enough was taking too much memory.

This commit is about removing the plt_auto_close trick and replace it by a proper memory management.

When the Graph() object is about to be destroyed, let's only close the associated figure making the close() call not racing with current renderings.

This patch was tested on a large rendering (7849 graphs) reduced the overall time from 70 minutes + crash (lack of memory) to 19 minutes with no crash.

Signed-off-by: Erwan Velu <e.velu@criteo.com>