hwbench: Add original configuration in result json
This commit add a dict into the result json representing the original
file configuration.

The parsed configuration is stored under the "config" key in the json
output.

As we use a rawconfigparser, we'll not have the general section saved
here but only each individual job merged with the general section.

This feature will be useful to detect if two results.json were made with
the same configuration.

Please note this is the raw content of the configuration file, there is
no interpretation of any configuration directive and neither the list of sub
jobs generated by the expanding.

A typical output looks like :
  "config": {
    "cpu": {
      "runtime": "60",
      "monitor": "all",
      "stressor_range": "auto",
      "stressor_range_scaling": "plus_1",
      "hosting_cpu_cores": "simple",
      "hosting_cpu_cores_scaling": "iterate",
      "engine_module_parameter_base": "",
      "engine": "stressng",
      "engine_module": "cpu",
      "engine_module_parameter": "matrixprod"
    }
  }

Signed-off-by: Erwan Velu <e.velu@criteo.com>