hwgraph: Support asymetric plots in Scaling graphs
The current code enforced all traces must be strictly identical
preventing comparing system with different number of cpu cores.

This commit is about adding the asymetrical support.

Comparing the list of benches to valid all traces are
identical is replaced by comparing the orignal configuration file.
hwbench store the config file in the "config" section of the
results.json. We'll use it to compare if all traces are similar in their
definition.

At rendering time, jobs will not have the same name and jobs count may
change. On the same job definition like :
	[global]
	runtime=120
	monitor=all
	engine=stressng
	hosting_cpu_cores=simple
	hosting_cpu_cores_scaling=iterate
	stressor_range=auto

	[cpu]
	engine_module=cpu
	engine_module_parameter=matrixprod,fft

This is expanded like the following :
	24 cores : 12 benchmarks
	32 cores : 14 benchmarks
	64 cores:  16 benchmarks

To ease the processing of this, two new functions are added in Trace():
- get_benches_by_job()
- get_benches_by_job_per_emp()

These helpers will precompute which jobs, per trace file are related to
the same engine module parameter run.

At rendering time, it will be important to keep in mind that each trace
file has a different numbers of points to plot. This is done by
adjusting the number of 'workers' for a given plot to match the number
of performance metrics we found during the performance metrics gathering.

This commit makes the scaling graphs accepting asymetrical plots,
another commit will take care of the individual ones.


Signed-off-by: Erwan Velu <e.velu@criteo.com>