hwgraph: Add asymetrical support for individual graphs
Individual graphs are made to compare each trace for each benchmark
using bar graphs.

The code was taking all jobs one after the other and compared them.

This is no more possible as the number of jobs may vary between traces
as the expanding could be different regarding the number of cores on a
system.

The logic was updated to consider the rendering per job (as per the
scaling graphs).

One issue to tacle was handling the non-existing values when one system
as much cores than the other.

The choosen solution is to initialized lists with 0 by default and
update the performance metric when one is found in a trace.

To achieve this:
- add_perf() receive a new parameter (index) to select which entry of
  the list is used instead of appending to it
- all aggregated data are computed once for all sub jobs on a given jobs
  and rendered in a single loop. This is improving rendering speed.

The rendering is also improved by adding minor lines behind the bars.

A specific graph is added to summarize the maximum results per run
allowing to compare, at the max perf, what all products are able to
deliver.


Signed-off-by: Erwan Velu <e.velu@criteo.com>