hwbench/stressng: provide homogenous skipping feature
Benchmarks are skipped under one of those two conditions:
 * current CPU does not support a given feature (ex AVX/VNNI)
 * stress-ng version is different to what we expect, breaking the
   parsing. Ideally this should be detected at the beginning during
   parameter validation, but we don't test the stress-ng version until
   after the first run. So we need to be in run_cmd() at least to have
   the correct version.

We already decided to do the skipping for memrate but not for stream,
which led to inconsistencies. Skipping will now happen in both cases.

In addition, we document what an "empty" result looks like and provided
a "skipped" key in the output for both cases.

Format of the results was not modified in order to keep hwgraph working.