hwbench: Ensure 'simple' helper will take all available cores
The current code was selecting cores on a modulo 16 like :
[1, 2, 3, 4, 8, 16, 32, 48, 64, 80, 96, 112, 128, 144, 160, 176, 192, 208, 224, 240, 256]

If a system has 36 cores (72 threads), the biggest value used will be
the closest in the above list, meaning the maximum number of cores will
not be tested.

This commit adds a rule where if the number of cores is not is this
list, it's automatically added to ensure a "all-cores" case will be
tested.

This commit also add an assert to ensure the number of benchmarks is
properly generated during testing.

An Intel 6240 output (lscpu, numactl) is added to test this case.

Signed-off-by: Erwan Velu <e.velu@criteo.com>