hwgraph: Add max_perf_per_core individual graph
The current code was generating a graph to compare the best performance of all traces.
Since traces might not have the same number of cores, comparing the
maximum performance is interesting.

That said, we miss one graph to compare the maximum performance but
divided by the number of cores.

This new graph will provide a good hint of the per-core performance when
the system reachs it's global maximal performance.

Signed-off-by: Erwan Velu <e.velu@criteo.com>