hwgraph: Adding dots in scaling graphs
This commit is adding dots for each measure in the scaling graph.
This is useful to help the visualization of the scaling used during the
benchmark. This is especially true when using 'simple' helper which
defines automatically the iteration steps regarding the CPU
architecture.

The 'o' marker makes a very little dot for each trace value.

Signed-off-by: Erwan Velu <e.velu@criteo.com>