hwbench: Adding BMC() class
This commit is about adding a BMC class to be called by vendors when
they need to interact with the bmc.

Two new methods are added to Vendor class :
- prepare()
  This method is called once the vendor matched.
  It creates the BMC object, could be custom per vendor.

- get_bmc() : Return the BMC object created by prepare()

The BMC class is a generic implementation using ipmitool.
For now, the main method is get_ip() to be overrided if needed by
vendors. The associated tests are added to ensure a proper parsing.

In the HPE case, ilorest is used to implement get_ip() as some products
like RL300 doesn't have any open interface anymore.
To get the BMC ip, an ILOREST class is added to manage the ilorest
behaviors : login/logout/...
An ILO class is created using ILOREST class as provider to extract the
BMC ip.

This first implementation allows to
- get the BMC IP when starting the monitoring. If not available, a fatal error is reported.
- connect to redfish and get the output of a query

Signed-off-by: Erwan Velu <e.velu@criteo.com>