hwbench: Add read_fans() to BMC Class
This commit is about adding the fans reporting.

As they are taken from the same Thermal endpoint, the code is a bit
refactored to make some functions more generic.

This commit is also adding the associated tests to ensure data is
properly parsed by the code.

The fans name are not adjusted.

Signed-off-by: Erwan Velu <e.velu@criteo.com>