hwbench: Add read_power_supplies() to BMC class
This commit is about adding a read_power_supplies() helper for the BMC class.

There is a default implementation that Dell uses.

For HPE, we have to consume the Oem extensions to collect the metric.
PSUs are renamed to match a common reporting with the generic
implementation.

This commit also add the unit testing to ensure the parsing is
consistent across products and vendors.

Signed-off-by: Erwan Velu <e.velu@criteo.com>