hwbench: Use dataclass instead of custom class
The previous implementation of MonitorMetric class was using a custom
logic to store metrics including value, name & unit. Some helpers were
added to manage comparison and string representations.

Since Python 3.7, dataclass decorator can be used to manage all this.

Dataclass defines for each property of the class if it should be part of
- the default constructor
- used an optional argument
- is part of the comparison when doing __eq__
- is part of the repr/str
etc..

Migrating to dataclass makes code shorter and more Pythonish.

This commit is also changing the way metrics are collected and stored.
Instead of having a single metric per MonitorMetric, 'value' is
replaced by a list of values.

Everytime the add() method is called, a new metric with the same
MonitorMetric type (unit/name) is added.

Once enough metrics are collected, the new compact() method can be
called to compute min/max/mean/stddev from 'values' member and
store them as a new entry if the associated lists.

This will greatly simplify how metrics will be processed in the monitoring
loops.

Note that MonitorMetric() keep an optional argument 'value' on the
constructor to shorten the tests by adding a single value in a single
line of code. Parameters order is changed explaining why some tests are
modified here.

This commit didn't get merged into previous commits to :
- reduce the work of changing to this data model
- to keep the history of why we changed to this data model


Signed-off-by: Erwan Velu <e.velu@criteo.com>