hwbench: Adding turbostat support
This commit adds turbostat metrics to the monitoring.

When Turbostat class is instantiated, a first run of turbotstat is done
to detect available features and ensure turbostat will work when needed.

When the monitoring loop starts, a turbostat is executed in background
via the run() method.

Just before restarting the next iteration, the output is decoded and
added to the CPU metrics.

This commit is also adding the turbostat mocking when parsing the
configuration file in tests :
- TestCommon.parse_config() creates a Monitoring object
- Monitoring instantiate Turbostat() if x86_64
- As the CPU mocking is x86_64, a Turbostat object is created
- Turbostat makes a dry run when instanciated to detect capabilites
- Turbostat.run() is mocked by hwbench/tests/parsing/turbostat/run
- TestParse.test_monitoring() tests if turbostat detects all required metrics

This commit is also updating the data model:
- Power metrics collected via the BMC are under PowerConsumption/BMC instead of PowerConsumption/Power

Turbostat metrics are split in two:
- cpu power metrics are located in PowerConsumption/CPU
- cpu frequency metrics are located in Freq/CPU

This v1 runs one turbostat per Monitoring.__monitor() inner loop which
isn't ideal but good enough to start.

Signed-off-by: Erwan Velu <e.velu@criteo.com>