hwgraph: Implement list sub-command
hwgraph requires a power metric in the trace description.
It's not an easy task to know what are the possible values.

This patch is about adding a 'list' command, which is exclusive with
'graph' command to list the power metrics from a trace file.

As the file format changed since the latest commits of hwbench, this
commit is also :
- moving imports to be at package level
- reuse monitoring classes from hwbench to keep the same data type
- add a list() method to Enum types
- add a list_str() method to Enum types
- add MonitoringMetadata Enum to avoid hardcoded strings
- add Bench().load_monitoring() to recreate MonitorMetric structures
- add MonitorMetric.load_from_dict() to repopulate a MonitorMetric from
  a dict loaded in the json output file.
- Add PowerCategories to avoid hardcoded strings

To achieve this feature, subparsers are added for each command.

Syntax looks like :

	python -m graph.hwgraph graph --traces dell/default/XXXXXXX dell/default/XXXXXXX

	python -m graph.hwgraph list --trace dell/default/8F7MQ0


Signed-off-by: Erwan Velu <e.velu@criteo.com>