hwbench/monitoring: Do not compact() after incomplete period
The compact() is done at every 'frequency' but an additional one was
added at the end of the loop.

As a consequence, if metrics were collected below the 'frequency' count,
this was creating measures with a different sample count and a different
running time.

So let's drop the compact after a monitoring loop and only compute it
once "frequency" is reached.

Signed-off-by: Erwan Velu <e.velu@criteo.com>