hwgraph: Rename --same-enclosure option to --same-chassis
This commit is about:
- fixing the chassis/enclosure semantic
- update the logic to support the new file format

The current code was using the enclosure semantic that is actually
chassis and chassis is replaced by server.

The existing code was graphing the chassis versus the sum of the
servers. As we have now two metrics for servers (SERVER and
SERVERINCHASSIS), let's plot the graph twice to get every metric plot
aside, unless it's pretty difficult to read the graph.

In addition this commit :
- add __eq__() PowerCategories to ease comparision
- load_monitoring() is called for each Bench() object created
- add get_time_interval(), get_component(), get_single_metric(), get_samples_count()
  these helpers will be reused in other part of the project to reach the
   internal datastructures. This might change a bit in later commits.
- compute the sample time resolution based on get_time_interval()
  instead of a 10sec harcoded value

Signed-off-by: Erwan Velu <e.velu@criteo.com>