stress-ng: Implement effective_runtime reporting
Sometimes stress-ng do not statisfy the timing constraints.

The command line requires a 30 sec run but the "real time" reported by
stress-ng is much shorter like < 20 sec.

stress-ng do not report any error because of a too short runs and
provides performance metrics, as for a valid run.

The performance reported can be seriously biased as the runtime is not
comparable with other runs that matched the expected duration.

This commit is about adding a new field name "effective_runtime" in the output metrics by
extracting the "real time" metric from stress-ng output.

To achieve that, a new StressSNG.stats_parse() method is added to match
the line reporting the stats of a stress-ng run.
This function is returning named fields to be used by any part of the
code that need some metrics of this type.
StressNG.parse_cmd() is updated to use it, removing the static
parsing of the output.

This metric will be used by hwgraph to detect invalid content and avoid
graphing non-comparable runs.

Tests are updated accordingly.


Signed-off-by: Erwan Velu <e.velu@criteo.com>