monitoring/turbostat: Pin turbostat to the last cpu core
Turbostat is spawned during the monitoring to collect cpu-based metric.
This tool is running in userpace and will be at some point running on
the same cores as the benchmark itself.

To get a more predictable behavior, let's stick turbostat to the last logical
cpu core :
- during scaling jobs, until we load all cores, turbostat will not
  compete with stress-ng
- when all cores are used, stress-ng will be sticked on a fixed core
  preventing kernel rescheduling to other cores

This removed the jitter messages from turbostat like :

	turbostat: cpu100 jitter 2461 575

Signed-off-by: Erwan Velu <e.velu@criteo.com>