hwgraph/trace: Explicit missing metric instead of crashing
Sometimes, the graph is crashing because traces are invalid, a typical
crash look like :

  File "/home/e.velu/Devel/hwbench/graph/trace.py", line 239, in add_perf
    value = self.get(perf)["sum_speed"]
            ~~~~~~~~~~~~~~^^^^^^^^^^^^^
TypeError: 'NoneType' object is not subscriptable

It's not that easy considering the number of benches and traces to
understand where the issue is located.

This commit is adding an explicit fault when such case arise. A typical
output looks like :

	Fatal: XXXXXXX/memrate_55: unable to find metric write8/sum_speed

Signed-off-by: Erwan Velu <e.velu@criteo.com>