hwgraph/trace: Report warning if benchmark didn't finish on time
For some reasons, some stress-ng instance ran in a shorter time than
expect. One of the main issue is that the reporting performance could be
very different to what it would have been if the job last the proper
amount of time.

It's important to warn about this issue as some graphs will report improper values.
This could be very misleading when reading the graphs.

This first version allows up to 1 sec of time difference.

For now, let's report this at graph time as a warning.
A typical output looks like :

	Individual: rendering 6 jobs
	XXXXXXX/cpu_39 didn't completed on time. Effective_runtime=28.09 vs 30 : delta=[1.91s; 6.37%]
	XXXXXXX/cpu_32 didn't completed on time. Effective_runtime=22.18 vs 30 : delta=[7.82s; 26.07%]
	XXXXXXX/cpu_41 didn't completed on time. Effective_runtime=16.21 vs 30 : delta=[13.79s; 45.97%]
	XXXXXXX/cpu_41 didn't completed on time. Effective_runtime=28.28 vs 30 : delta=[1.72s; 5.73%]

Signed-off-by: Erwan Velu <e.velu@criteo.com>