hwbench: default stress-ng to version 0.17.4
The current code was supporting several versions of stress-ng and
enabled some features based on the release.

The expected version to get most of the feature was 0.16.4.

This release is affected by a couple of bugs including some vnni issues
and some timing management troubles leading to unprecise results.

This commit is about enforcing 0.17.4 as a minimal version:
- all tests prior 0.17.4 are dropped
- tests are defaulted to 0.17.4 output
- all code testing release is dropped
- all engines_modules enforce a 0.17.4 as a minimum
- -quiet option is dropped to get a constant output for all command
- -metrics replace -metrics-brief option to get the expected output

Signed-off-by: Erwan Velu <e.velu@criteo.com>