hwgraph: Adding error bars on power scaling graph
When plotting scaling graph, one is about the power consumption.
It would be interesting to see the distribution of the power consumption
on the graph itself.

So let's add error bars on each measure.

To achieve that, values are computed in add_perf() if needed.

Colors are enforced on this graph to ensure a good match between the
error bar color (red) and the data line.

Signed-off-by: Erwan Velu <e.velu@criteo.com>