hwgraph: Fix minor ticks plotting
The current code was buggy and generated unreadable graphs with larger number of points to plot.

The first mistake is the number of points_to_plot was computed on
next(iter(sum_serie_to_plot) which is the name of the metric ... so the
len() was measuring the length() of the metric name instead of counting
the number of points.

The second issue is that plt.minorticks_on() reset the minor locator to
a different value than the one fixed earlier in the code.
This value was not a modulo of the monitoring time creating unaligned
points and minor grid. So this commit is doing a single set of set_minor_locator().

The last change is to make chassis graphs using numbers related to the
actual time interval instead of hardcoded values.

This commits makes clearer graphs to read.

Signed-off-by: Erwan Velu <e.velu@criteo.com>