hwgraph: Force y2 ticks to match y1
When using two y axis, as per the server graphs like fans, the main y
axis is driving the grid while the second one is unaligned to it.

It result in a very difficult reading of values on the y2 axis as the y1
grid is not aligned with the y2 axis.

This commit ensure the y2 axis is having the same number of ticks as y1
aligning the y2 ticks on y1. y2 values are no more rounded but at least
its easier to project any measure to it.

Signed-off-by: Erwan Velu <e.velu@criteo.com>