hwbench: Add sync_start option
When running the same bench on a multi-node chassis, it could be
interesting to get all jobs to be synchronized. Even if you spawn all
commands at the same time, the time to set up the BMC connection may
vary from one node to another, that's why it's important to resync
systems once ready.

For now, there is no external management to perform this task so let's
try to get a simple implementation of such feature.

This commit is adding a new option : sync_start

It has a default value set to "none" to keep the same behavior as today.
If sync_start=time, it enables a sync behavior :
- nodes have to be time synced with NTP
- each node consider the next meeting point by finding the next O'clock
  minute.
- If we too close for the next meeting point, it is deferred by another
  meeting to ensure all nodes will have time to join the meeting point.

This is a pretty simple algorithm but prooved to be OK when testing
multi-node chassis servers.

/!\ Please note that NTP is really mandatory to use this feature /!\


Signed-off-by: Erwan Velu <e.velu@criteo.com>