hwbench: Add auto mode to spike engine
The current spike engines requires to define the low and high values to
function.

This patch is about adding an automatic mode to avoid interating around
values to detect a particular behavior.

The algorithm looks like the following :
- the sum of the fan speed is calibrated to reach a low value over 30
  seconds.
- server is loaded with stress-ng matrixprod workload.
  the number of logical cpu cores to be used is still defined by the
  usual 'hosting_cpu_cores' and 'stressor_range' parameters.
- once the fans increased their speed by <x>%, the cpu load is stopped
- a loop waits the fans to return to their initial fan speed
- if we have enough time, this loop is restarted

The usual runtime parameter for this case, defines the overal time to
perform loops of fan spiking. Regarding fan's behavior, the number of
iteration within the runtime may vary.

At the end of the job, two metrics are added to the results.
- time_to_high: a list of floats reporting the time needed to reach the
  fan limit before stopping the cpu workload
- time_to_low: a list of floats reporting the time need to return to the
  initial low fan speed.

By default, the 'auto' mode considers the cpu load must be stopped once
fans speed up by 5%. This can be overidded by padding a number after the
auto mode directive like in the following example showing a 'stop if
fans speed up by 10%' example.

	engine_module_parameter_base=auto:10

This mode works pretty well on chassis that have a single server inside.
A customized stop value must be adjusted regarding the system : number
of fans, units and thermal algorithm's sensisivity.
Note some servers reports their fans in RPM and others in percent. This
code is computing the percentage on top the sum of these metrics,
implicating the precision may vary between systems.

This algorithm is a bit less efficient on multi-node chassis because
each node, even if synchronized, is not getting updates from the BMC at
the same time :
- each nodes reads on a time basis the BMC fans' speed
- BMC on its side compute an average of fan speed to be updated every
  _t_ seconds
- but all BMCs on the same chassis are not reading, and computing, the
  average _t_ seconds at the same time
- each servers has a different view of the same chassis : one can see
  the sum of fans at X while the next server see them at X+-10%.
All all servers cannot be synchronized on the value their read at a
given time, some node may decide to stop or start the cpu load which
could be a contradictory state seen by others nodes, influencing them in
their next decision.
Some nodes may not see the fans reaching the low values (in a descent)
because others decided to load them again at the same time.
Even with this possible issue, it doesn't occurs all time, the job will
output possible time_to_high and time_to_low values to be used manually
for fine grained testing.

This code still has some printf enabled to validate the behavior on more
machines.

Signed-off-by: Erwan Velu <e.velu@criteo.com>