hwgraph/chassis: Do not crash if missing samples
In rare condition (tm), some traces might miss one metric, probably
because of a very little timing difference, generating a crash at render
time.

The current logic consider that all traces must have the same number of
monitoring samples but it might happen that one is missing. This led to
a hwgraph crash, preventing all traces to be drawn.

This commit is about testing if a given trace & job has the proper
number of metrics. If not, a warning is reported and value is considered
to be 0 watts.

The first implementation was trying not to set any value but it was
preventing generating the graph as matplotlib requires for this graph
having the same number of trace points. To avoid this failure and to
ensure the other graphs can be done, let's put the value to 0 and warn
the user about this issue.

A typical output looks like :
   environment: All traces are from the same chassis (XXXXXXX), enabling --same-chassis feature
   Warning: XXXXXXX/qsort_63: Missing sample 11, considering 0 watts.

Signed-off-by: Erwan Velu <e.velu@criteo.com>